Add export functionality
It's nice to have the feature of exporting data as a .csv or .xlsx.

This also means they can be imported too.

This feature means if the app is run locally the data is not lost, and also can be easily shared with your line manager as an Excel file.